# api-docs.json
{"openapi":"3.1.0","info":{"title":"Pedidos API - Treinamento Java Avançado:Formação Arquiteto","description":"Documentação da API do projeto","contact":{"name":"COTI Informática","email":"contato@cotiinformatica.com.br"},"version":"v1"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"Pedidos","description":"Endpoints para gerenciamento de pedidos."}],"paths":{"/api/v1/pedidos/{id}":{"get":{"tags":["Pedidos"],"summary":"Obter um pedido","description":"Endpoint para obter um pedido existente.","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PedidoResponse"}}}}}},"put":{"tags":["Pedidos"],"summary":"Atualizar um pedido","description":"Endpoint para atualizar um pedido existente.","operationId":"put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PedidoRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PedidoResponse"}}}}}},"delete":{"tags":["Pedidos"],"summary":"Excluir um pedido","description":"Endpoint para excluir um pedido existente.","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PedidoResponse"}}}}}}},"/api/v1/pedidos":{"get":{"tags":["Pedidos"],"summary":"Obter uma lista de pedidos","description":"Endpoint para obeter uma lista de pedidos existente.","operationId":"get","parameters":[{"name":"pagina","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"qtdRegistros","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":25}},{"name":"campoOrdenacao","in":"query","required":false,"schema":{"type":"string","default":"dataPedido"}},{"name":"tipoOrdenacao","in":"query","required":false,"schema":{"type":"string","default":"asc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagePedidoResponse"}}}}}},"post":{"tags":["Pedidos"],"summary":"Inserir um pedido","description":"Endpoint para inserir um novo pedido.","operationId":"post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PedidoRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PedidoResponse"}}}}}}}},"components":{"schemas":{"PedidoRequest":{"type":"object","properties":{"valor":{"type":"number","format":"double","minimum":0.01},"cliente":{"type":"string","maxLength":100,"minLength":8},"detalhes":{"type":"string","maxLength":250,"minLength":8}},"required":["cliente","detalhes","valor"]},"PedidoResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"dataPedido":{"type":"string","format":"date-time"},"valor":{"type":"number","format":"double"},"cliente":{"type":"string"},"detalhes":{"type":"string"},"status":{"type":"string"}}},"PagePedidoResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/PedidoResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"unsorted":{"type":"boolean"},"sorted":{"type":"boolean"}}}}}}